(Metrics) hsp imprime results

+++ Metrics/scripts/human_scanpath_prediction.py
@@ -9,7 +9,7 @@ import shutil
 import numba
 import importlib
 
-""" Computes human scanpath prediction on the visual search models for a given dataset. 
+""" Computes Human Scanpath Prediction on the visual search models for a given dataset. 
     See "Kümmerer, M. & Bethge, M. (2021), State-of-the-Art in Human Scanpath Prediction" for more information.
     The methods for computing AUC and NSS were taken from https://github.com/matthias-k/pysaliency/blob/master/pysaliency/metrics.py
 """
@@ -34,14 +34,14 @@ class HumanScanpathPrediction:
 
         average_results_per_image = utils.load_dict_from_json(model_average_file)
         if average_results_per_image:
-            print('[Scanpath prediction] Found previously computed results for ' + model_name)
+            print('[Human Scanpath Prediction] Found previously computed results for ' + model_name)
         else:
             for subject in human_scanpaths_files:
                 subject_number = subject[4:6]
 
                 if not self.subject_already_processed(subject, subject_number, model_output_path):
                     model = importlib.import_module(self.models_dir + '.' + model_name + '.main')
-                    print('[Scanpath prediction] Running ' + model_name + ' on ' + self.dataset_name + ' dataset using subject ' + subject_number + ' scanpaths')
+                    print('[Human Scanpath Prediction] Running ' + model_name + ' on ' + self.dataset_name + ' dataset using subject ' + subject_number + ' scanpaths')
                     model.main(self.dataset_name, int(subject_number))
             
             average_results_per_image = self.average_results(model_output_path)
@@ -54,22 +54,22 @@ class HumanScanpathPrediction:
         subject_scanpath_file      = path.join(self.human_scanpaths_dir, subject_file)
         subject_predictions_file   = path.join(subjects_predictions_path, 'subject_' + subject_number + '_results.json')
         if utils.is_contained_in(subject_scanpath_file, subject_predictions_file):
-            print('[Scanpath prediction] Found previously computed results for subject ' + subject_number)
+            print('[Human Scanpath Prediction] Found previously computed results for subject ' + subject_number)
             return True
         
         return False
     
     def compute_model_mean(self, average_results_per_image, model_name):
         """ Get the average across all images for a given model in a given dataset """
-        self.models_results[model_name] = {'Scanpath_prediction': {'AUC': 0, 'NSS': 0, 'IG': 0}}
+        self.models_results[model_name] = {'Human_scanpath_prediction': {'AUC': 0, 'NSS': 0, 'IG': 0}}
 
         # Get subsample for COCOSearch18 dataset
         number_of_images            = min(len(average_results_per_image), self.number_of_images)
         results_per_image_subsample = utils.get_random_subset(average_results_per_image, size=number_of_images)
         for image_name in results_per_image_subsample:
-            self.models_results[model_name]['Scanpath_prediction']['AUC'] += results_per_image_subsample[image_name]['AUC'] / number_of_images
-            self.models_results[model_name]['Scanpath_prediction']['NSS'] += results_per_image_subsample[image_name]['NSS'] / number_of_images
-            self.models_results[model_name]['Scanpath_prediction']['IG']  += results_per_image_subsample[image_name]['IG'] / number_of_images
+            self.models_results[model_name]['Human_scanpath_prediction']['AUC'] += results_per_image_subsample[image_name]['AUC'] / number_of_images
+            self.models_results[model_name]['Human_scanpath_prediction']['NSS'] += results_per_image_subsample[image_name]['NSS'] / number_of_images
+            self.models_results[model_name]['Human_scanpath_prediction']['IG']  += results_per_image_subsample[image_name]['IG'] / number_of_images
     
     def average_results(self, model_output_path):
         """ Get the average of all subjects for each image """
@@ -110,11 +110,19 @@ class HumanScanpathPrediction:
 
         utils.save_to_json(dataset_metrics_file, dataset_metrics)
 
+    def print_results():
+        if self.null_object: return
+
+        print('[Human Scanpath Prediction] ' + self.dataset_name + ' dataset: ')
+        for model in self.models_results:
+            model_metrics = self.models_results[model]
+            print('(' + model + ') AUC:' + model_metrics['AUC'] + '; NSS: ' + model_metrics['NSS'])
+
 def save_scanpath_prediction_metrics(subject_scanpath, image_name, output_path):
     """ After creating the probability maps for each fixation in a given human subject's scanpath, visual search models call this method """
     probability_maps_path = path.join(output_path, path.join('probability_maps', image_name[:-4]))
     if not path.exists(probability_maps_path):
-        print('[Scanpath prediction] No probability maps found for ' + image_name)
+        print('[Human Scanpath Prediction] No probability maps found for ' + image_name)
         return
     probability_maps = listdir(probability_maps_path)
 
@@ -140,7 +148,7 @@ def save_scanpath_prediction_metrics(subject_scanpath, image_name, output_path):
     utils.save_to_json(file_path, model_subject_metrics)
 
     # Clean up probability maps if their size is too big
-    if utils.dir_is_too_heavy(probability_maps_path):
+    if not (image_name in ['grayscale_44_oliva.jpg', 'grayscale_6_airbnb.jpg', 'grayscale_1_kitchen.jpg'] and subject == 'subject_01') and utils.dir_is_too_heavy(probability_maps_path):
         shutil.rmtree(probability_maps_path)
 
 def compute_metrics(probability_map, human_fixation_y, human_fixation_x):
